requirements.txt
# ============================================================
# BoneQuest v2 — Python Dependencies
# ============================================================

# Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.18
aiofiles==24.1.0
pydantic==2.10.0

# Database
sqlalchemy==2.0.25

# External APIs
groq==0.13.0
pdfplumber==0.11.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Image Processing
Pillow==10.4.0

# Logging
structlog==24.4.0
